Extract BaseMetaTestsTest
