+++ src/test/java/integration/ModesAndActionsTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import static integration.CodeChecksTest.codeCheck;
-import static integration.LibraryMetaTestsTest.failedMetaTest;
+import static integration.BaseMetaTestsTest.failedMetaTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class ModesAndActionsTest extends IntegrationTestBase {
